requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
altair @ file:///croot/altair_1743016723624/work
annotated-types==0.7.0
anyio==4.10.0
attrs==25.3.0
awscli==1.42.33
backoff==2.2.1
beautifulsoup4==4.13.5
blinker @ file:///croot/blinker_1737448726027/work
boto3==1.40.33
botocore==1.40.33
Bottleneck==1.6.0
Brotli @ file:///croot/brotli-split_1736182456865/work
brotlicffi @ file:///croot/brotlicffi_1736182461069/work
cachetools==5.5.2
certifi @ file:///croot/certifi_1754570635119/work/certifi
cffi==2.0.0
charset-normalizer==3.4.3
click @ file:///croot/click_1754477934456/work
cohere==5.18.0
colorama==0.4.6
contourpy @ file:///croot/contourpy_1732540045555/work
cryptography==45.0.7
cycler @ file:///tmp/build/80754af9/cycler_1637851556182/work
dataclasses-json==0.6.7
docstring_parser==0.17.0
docutils==0.19
emoji==2.14.1
et_xmlfile==2.0.0
faiss-cpu==1.12.0
fastavro==1.12.0
filelock==3.19.1
filetype==1.2.0
fonttools @ file:///croot/fonttools_1737039080035/work
frozenlist==1.7.0
fsspec==2025.9.0
gitdb @ file:///croot/gitdb_1753362118172/work
GitPython @ file:///croot/gitpython_1720455034730/work
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-aiplatform==1.112.0
google-cloud-bigquery==3.37.0
google-cloud-core==2.4.3
google-cloud-resource-manager==1.14.2
google-cloud-storage==2.19.0
google-crc32c==1.7.1
google-genai==1.36.0
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
greenlet==3.2.4
grpc-google-iam-v1==0.14.2
grpcio==1.74.0
grpcio-status==1.74.0
h11==0.16.0
hf-xet==1.1.9
html5lib==1.1
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.34.4
idna @ file:///croot/idna_1714398848350/work
Jinja2 @ file:///croot/jinja2_1741710844255/work
jmespath==1.0.1
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema @ file:///croot/jsonschema_1753206402902/work
jsonschema-specifications @ file:///work/perseverance-python-buildout/croot/jsonschema-specifications_1701731664072/work
kiwisolver @ file:///croot/kiwisolver_1737039087198/work
langchain==0.3.27
langchain-aws==0.2.33
langchain-cohere==0.4.5
langchain-community==0.3.29
langchain-core==0.3.76
langchain-google-vertexai==2.0.28
langchain-tavily==0.2.11
langchain-text-splitters==0.3.11
langdetect==1.0.9
langgraph==0.6.7
langgraph-checkpoint==2.1.1
langgraph-prebuilt==0.6.4
langgraph-sdk==0.2.6
langsmith==0.4.27
lxml==6.0.1
Markdown==3.9
MarkupSafe @ file:///croot/markupsafe_1738584038848/work
marshmallow==3.26.1
matplotlib==3.10.6
msoffcrypto-tool==5.4.2
multidict==6.6.4
mypy_extensions==1.1.0
narwhals @ file:///croot/narwhals_1742845957875/work
networkx==3.5
nltk==3.9.1
numexpr==2.12.1
numpy @ file:///croot/numpy_and_numpy_base_1725470312869/work/dist/numpy-2.0.1-cp312-cp312-linux_x86_64.whl#sha256=0a2920303d36e9da3063f2d20885cb28d54fe6af2e1bee9bb30b48855a54664e
olefile==0.47
openpyxl==3.1.5
orjson==3.11.3
ormsgpack==1.10.0
packaging @ file:///croot/packaging_1753775359799/work
pandas @ file:///croot/pandas_1756466404410/work/dist/pandas-2.3.2-cp312-cp312-linux_x86_64.whl#sha256=91f9e704d6b71af3975875fd6b728a0e7b6474d918af2f04b67d23518eb52cc6
patsy @ file:///croot/patsy_1738159930729/work
pillow @ file:///croot/pillow_1752524526621/work
propcache==0.3.2
proto-plus==1.26.1
protobuf @ file:///croot/protobuf_1742419705443/work/bazel-bin/python/dist/protobuf-5.29.3-cp312-abi3-linux_x86_64.whl#sha256=b288453436273a46653983ade721c0c62cfa5ef6c012b110674c369d25536d88
psutil==7.0.0
pyarrow @ file:///croot/pyarrow_1756839868845/work/python
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser @ file:///home/task_175749564075091/conda-bld/pycparser_1757496039511/work
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydeck @ file:///croot/pydeck_1738340956155/work
pyparsing @ file:///croot/pyparsing_1731445506121/work
pypdf==6.0.0
PySide6==6.9.2
PySocks @ file:///work/perseverance-python-buildout/croot/pysocks_1698845478203/work
python-dateutil @ file:///croot/python-dateutil_1716495738603/work
python-dotenv==1.1.1
python-iso639==2025.2.18
python-magic==0.4.27
python-oxmsg==0.0.2
pytz @ file:///croot/pytz_1752135852232/work
PyYAML @ file:///croot/pyyaml_1728657952215/work
RapidFuzz==3.14.1
referencing @ file:///work/perseverance-python-buildout/croot/referencing_1701731622327/work
regex==2025.9.1
requests @ file:///croot/requests_1756709366904/work
requests-toolbelt==1.0.0
rpds-py @ file:///croot/rpds-py_1736541261634/work
rsa==4.7.2
s3transfer==0.14.0
scipy @ file:///home/task_175731568666045/conda-bld/scipy_1757315702136/work/dist/scipy-1.16.1-cp312-cp312-linux_x86_64.whl#sha256=22c5322de03d8db7e313c7138ba8c5f12a0e6b1e9c194ac88d0820c78cf3e713
seaborn @ file:///home/conda/feedstock_root/build_artifacts/seaborn-split_1733730015268/work
setuptools==78.1.1
shapely==2.1.1
shiboken6==6.9.2
six @ file:///croot/six_1744271502820/work
smmap @ file:///croot/smmap_1752748938695/work
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.43
statsmodels @ file:///croot/statsmodels_1753363813252/work
streamlit @ file:///home/conda/feedstock_root/build_artifacts/streamlit_1756565351169/work
tavily-python==0.7.12
tenacity==9.1.2
tiktoken==0.11.0
tokenizers==0.22.0
toml @ file:///tmp/build/80754af9/toml_1616166611790/work
tornado @ file:///croot/tornado_1748956929273/work
tqdm==4.67.1
types-PyYAML==6.0.12.20250822
types-requests==2.32.4.20250809
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions @ file:///croot/typing_extensions_1756280817316/work
tzdata @ file:///croot/python-tzdata_1746123641790/work
unicodedata2 @ file:///croot/unicodedata2_1736541023050/work
unstructured==0.18.14
unstructured-client==0.42.3
urllib3 @ file:///croot/urllib3_1750775463400/work
validators==0.35.0
watchdog @ file:///croot/watchdog_1740409902579/work
webencodings==0.5.1
websockets==15.0.1
wheel==0.45.1
wrapt==1.17.3
xlrd==2.0.2
xxhash==3.5.0
yarl==1.20.1
zstandard==0.24.0
